% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokedex}
\alias{pokedex}
\title{Pokedex of Pokemon GO}
\format{
A nested list containing 151 sublists with up to 17 list elements:
\describe{
  \item{id}{integer, Identification number}
  \item{num}{character, Pokemon number in official Pokedex}
  \item{name}{character, Pokemon name}
  \item{img}{character, URL to png image of the Pokemon}
  \item{type}{character, Pokemon type}
  \item{height}{character, Pokemon height}
  \item{weight}{character, Pokemon weight}
  \item{candy}{character, type of candy used to evolve Pokemon or given when transfered}
  \item{candy_count}{integer, amount of candies required to evolve}
  \item{egg}{character, travel distance to hatch the egg}
  \item{spawn_change}{numeric, spawn change percentage}
  \item{avg_spawns}{integer, number of spawns per 10.000 spawns}
  \item{spawn_time}{character, local time at which spawns are most active}
  \item{multipliers}{numeric, multiplier of Combat Power (CP) after evolution}
  \item{weakness}{character, types of Pokemon this Pokemon is weak to}
  \item{next_evolution}{list, numbers (\code{num}) and names (\code{name}) with successive evolutions}
  \item{prev_evolution}{list, numbers (\code{num}) and names (\code{name}) with previous evolutions}
}
}
\source{
\href{https://raw.githubusercontent.com/Biuni/PokemonGO-Pokedex/master/pokedex.json}{PokemonGO-Pokedex}
}
\usage{
pokedex
}
\description{
A nested list containing property values of each Pokemon in Pokemon GO. The data is available in JSON format from GitHub (credits to Gianluca Bonifazi).
}
\keyword{datasets}
